requirements.txt
numpy==1.21.4
opencv-python==4.5.4.60
python-dotenv==0.19.2
